Make GpuDecimalMultiplyBase shim specific with different parental heirarchy
