+++ sql-plugin/src/main/311until340-non330db/scala/org/apache/spark/sql/rapids/arithmetic.scala
@@ -18,10 +18,12 @@ package org.apache.spark.sql.rapids
 
 import ai.rapids.cudf.BinaryOp
 import com.nvidia.spark.rapids._
+import com.nvidia.spark.rapids.shims.ShimExpression
 
 import org.apache.spark.sql.catalyst.expressions.{Expression, NullIntolerant}
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types._
+import org.apache.spark.sql.vectorized.ColumnarBatch
 
 abstract class CudfBinaryArithmetic extends CudfBinaryOperator with NullIntolerant {
   override def dataType: DataType = left.dataType
@@ -50,16 +52,24 @@ case class GpuDecimalDivide(
 case class GpuDecimalMultiply(
     left: Expression,
     right: Expression,
-    override val dataType: DecimalType,
+    dataType: DecimalType,
     useLongMultiply: Boolean = false,
-    failOnError: Boolean = SQLConf.get.ansiEnabled)
-    extends GpuDecimalMultiplyBase(left, right, dataType, useLongMultiply, failOnError) {
-  override def inputType: AbstractDataType = DecimalType
+    failOnError: Boolean = SQLConf.get.ansiEnabled) extends ShimExpression with GpuExpression
+    with GpuDecimalMultiplyBase {
 
-  override def symbol: String = "*"
+  override def sql: String = s"(${left.sql} * ${right.sql})"
 
-  // We aren't using this
-  override def binaryOp: BinaryOp = BinaryOp.MUL
+  override def columnarEval(batch: ColumnarBatch): Any = {
+    if (useLongMultiply) {
+      longMultiply(batch)
+    } else {
+      regularMultiply(batch)
+    }
+  }
+
+  override def nullable: Boolean = left.nullable || right.nullable
+
+  override def children: Seq[Expression] = Seq(left, right)
 }
 
 case class GpuAdd(
