+++ tests/src/test/scala/org/apache/spark/sql/rapids/catalyst/expressions/GpuEquivalentExpressionsSuite.scala
@@ -539,8 +539,7 @@ class GpuEquivalentExpressionsSuite extends FunSuite with Logging {
     val price: AttributeReference = AttributeReference("price", DecimalType(7, 2))()
     val inputAttrs = AttributeSeq(Seq(customer, quantity, price))
 
-    val product = GpuDecimalMultiply(
-      GpuCast(quantity, DecimalType(12, 2)), GpuCast(price, DecimalType(12, 2)), DecimalType(18,2))
+    val product = GpuDecimalMultiply(quantity, price, DecimalType(18,2))
     val nullCheck = GpuIsNull(product)
     val castProduct = GpuCast(product, DecimalType(28,2))
     val extract0 = GpuExtractChunk32(castProduct, 0, true)
